requirements.txt
python-dotenv
pypdf

fastapi==0.115.8
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
groq==0.9.0
python-telegram-bot==21.6

